dual_model_divergence_project/requirements.txt
openai>=1.0.0
anthropic>=0.34.0
python-dotenv>=1.0.0
pytest>=8.0.0

